Implemented AsteroidSpawner class
